修改掉之前的语法错误，能够走通编译，打印处理class里面的数据

+++ classfile/constant_utf8_info.go
@@ -5,7 +5,7 @@ type ConstantUtf8Info struct {
 }
 
 func (self *ConstantUtf8Info) readInfo(reader *ClassReader) {
-	length := reader.readUint16()
+	length := uint32(reader.readUint16())
 	bytes := reader.readBytes(length)
 	self.value = decodeMUtf8(bytes)
 }
